refactor: Domain 단위 파일 구조 변경
