+++ src/main/java/usay/app/domain/agent/entity/Agent.java
@@ -1,46 +1,46 @@
-package usay.app.common.domain;
+package usay.app.domain.agent.entity;
 
 import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
-import jakarta.persistence.EnumType;
-import jakarta.persistence.Enumerated;
 import jakarta.persistence.GeneratedValue;
 import jakarta.persistence.GenerationType;
 import jakarta.persistence.Id;
-import jakarta.persistence.JoinColumn;
-import jakarta.persistence.ManyToOne;
+import jakarta.persistence.OneToMany;
+import java.util.List;
 import lombok.AccessLevel;
 import lombok.Builder;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.NonNull;
 import usay.app.common.entity.BaseEntity;
+import usay.app.domain.room.entity.Room;
 
 @Getter
-@Entity
 @NoArgsConstructor(access = AccessLevel.PROTECTED)
-public class Room extends BaseEntity {
+@Entity
+public class Agent extends BaseEntity {
 
 	@Id
 	@GeneratedValue(strategy = GenerationType.IDENTITY)
 	private Long id;
 
 	@NonNull
-	@Column
 	private String name;
 
-	@Column
-	@Enumerated(EnumType.STRING)
-	private RoomStatus status;
+	@NonNull
+	private Integer difficulty;
 
-	@ManyToOne
-	@JoinColumn(name = "agent_id")
-	private Agent agent;
+	@NonNull
+	@Column(columnDefinition = "TEXT")
+	private String about;
+
+	@OneToMany(mappedBy = "agent", orphanRemoval = true)
+	private List<Room> rooms;
 
 	@Builder
-	public Room(String name, RoomStatus status, Agent agent) {
+	public Agent(String name, Integer difficulty, String about) {
 		this.name = name;
-		this.status = status;
-		this.agent = agent;
+		this.difficulty = difficulty;
+		this.about = about;
 	}
 }
