+++ src/main/java/usay/app/domain/agent/repository/AgentRepository.java
@@ -1,7 +1,7 @@
-package usay.app.common;
+package usay.app.domain.agent.repository;
 
 import org.springframework.data.jpa.repository.JpaRepository;
-import usay.app.common.domain.Agent;
+import usay.app.domain.agent.entity.Agent;
 
 
 public interface AgentRepository extends JpaRepository<Agent, Long> {
