+++ src/main/java/usay/app/domain/chat/repository/ChatRepository.java
@@ -1,4 +1,4 @@
-package usay.app.common.domain.chat.repository;
+package usay.app.domain.chat.repository;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 import usay.app.common.domain.Chat;
