+++ src/main/java/usay/app/domain/member/repository/MemberRepository.java
@@ -1,6 +1,6 @@
-package usay.app.common;
+package usay.app.domain.member.repository;
 
 import org.springframework.data.jpa.repository.JpaRepository;
-import usay.app.common.domain.Member;
+import usay.app.domain.member.entity.Member;
 
 public interface MemberRepository extends JpaRepository<Member, Long> {}
